+++ python/hello-world/src/hello_world/constraints.py
@@ -5,7 +5,7 @@ from datetime import time
 from .domain import Lesson
 
 @constraint_provider
-def school_timetabling_constraints(constraint_factory: ConstraintFactory):
+def define_constraints(constraint_factory: ConstraintFactory):
     return [
         # Hard constraints
         room_conflict(constraint_factory),
