+++ python/hello-world/tests/test_constraints.py
@@ -2,7 +2,7 @@ from timefold.solver.test import ConstraintVerifier
 from datetime import time
 
 from hello_world.domain import Timetable, Lesson, Room, Timeslot
-from hello_world.constraints import (school_timetabling_constraints, room_conflict,
+from hello_world.constraints import (define_constraints, room_conflict,
                                      teacher_conflict, student_group_conflict,
                                      teacher_room_stability, teacher_time_efficiency,
                                      student_group_subject_variety)
@@ -14,7 +14,7 @@ TIMESLOT2 = Timeslot("TUESDAY", time(12, 0), time(13, 0))
 TIMESLOT3 = Timeslot("TUESDAY", time(13, 0), time(14, 0))
 TIMESLOT4 = Timeslot("TUESDAY", time(15, 0), time(16, 0))
 
-constraint_verifier = ConstraintVerifier.build(school_timetabling_constraints, Timetable, Lesson)
+constraint_verifier = ConstraintVerifier.build(define_constraints, Timetable, Lesson)
 
 def test_room_conflict():
     first_lesson = Lesson("1", "Subject1", "Teacher1", "Group1", TIMESLOT1, ROOM1)
