+++ python/school-timetabling/src/school_timetabling/routes.py
@@ -6,7 +6,7 @@ from timefold.solver.config import (SolverConfig, ScoreDirectorFactoryConfig,
 from typing import Annotated
 
 from .domain import Timetable, Lesson, Room, Timeslot, ConstraintAnalysisDTO, MatchAnalysisDTO
-from .constraints import school_timetabling_constraints
+from .constraints import define_constraints
 from .demo_data import DemoData, generate_demo_data
 
 
@@ -14,7 +14,7 @@ solver_config = SolverConfig(
     solution_class=Timetable,
     entity_class_list=[Lesson],
     score_director_factory_config=ScoreDirectorFactoryConfig(
-        constraint_provider_function=school_timetabling_constraints
+        constraint_provider_function=define_constraints
     ),
     termination_config=TerminationConfig(
         spent_limit=Duration(seconds=30)
