+++ python/school-timetabling/tests/test_feasible.py
@@ -1,5 +1,5 @@
 from school_timetabling.routes import app
-from school_timetabling.domain import Timetable, Room, Timeslot
+from school_timetabling.domain import *
 
 from fastapi.testclient import TestClient
 from time import sleep
